# docs/api/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Real-Time Analytics Pipeline API",
    "version": "1.0.0",
    "description": "Enterprise-grade real-time analytics pipeline API. Supports event ingestion, analytics queries, dashboard management, and real-time WebSocket streaming.",
    "contact": {
      "name": "Engineering Team",
      "email": "engineering@example.com"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:8000",
      "description": "Local development"
    },
    {
      "url": "https://api.example.com",
      "description": "Production"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": { "type": "string", "example": "VALIDATION_ERROR" },
              "message": { "type": "string", "example": "Validation failed" },
              "details": { "type": "object", "example": {} }
            },
            "required": ["code", "message"]
          }
        },
        "required": ["error"]
      },
...
      "PaginatedResponse": {
        "type": "object",
        "properties": {
          "items": { "type": "array" },
          "total": { "type": "integer" },
          "page": { "type": "integer" },
          "page_size": { "type": "integer" },
          "total_pages": { "type": "integer" },
          "has_next": { "type": "boolean" },
          "has_previous": { "type": "boolean" }
        }
      },
...
      "WidgetCreateRequest": {
        "type": "object",
        "required": ["title", "widget_type", "metric_name", "position"],
        "properties": {
          "title": { "type": "string", "minLength": 1, "maxLength": 255 },
          "widget_type": { "type": "string", "enum": ["timeseries", "counter", "pie_chart", "table", "heatmap"] },
          "metric_name": { "type": "string" },
          "config": { "type": "object" },
          "position": { "type": "integer", "minimum": 0 },
          "width": { "type": "integer", "minimum": 1, "maximum": 12, "default": 6 },
          "height": { "type": "integer", "minimum": 1, "maximum": 12, "default": 4 }
        }
      },
      "DashboardCreateRequest": {
        "type": "object",
        "required": ["name"],
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string", "default": "" },
          "widgets": { "type": "array", "items": { "$ref": "#/components/schemas/WidgetCreateRequest" } },
          "is_public": { "type": "boolean", "default": false }
        }
      },
      "DashboardUpdateRequest": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "minLength": 1, "maxLength": 255 },
          "description": { "type": "string" },
          "widgets": { "type": "array", "items": { "$ref": "#/components/schemas/WidgetCreateRequest" } },
          "is_public": { "type": "boolean" }
        }
      },
      "WidgetResponse": {
        "type": "object",
        "properties": {
          "widget_id": { "type": "string", "format": "uuid" },
          "title": { "type": "string" },
          "widget_type": { "type": "string" },
          "metric_name": { "type": "string" },
          "config": { "type": "object" },
          "position": { "type": "integer" },
          "width": { "type": "integer" },
          "height": { "type": "integer" }
        }
      },
      "DashboardResponse": {
        "type": "object",
        "properties": {
          "dashboard_id": { "type": "string", "format": "uuid" },
          "name": { "type": "string" },
          "description": { "type": "string" },
          "widgets": { "type": "array", "items": { "$ref": "#/components/schemas/WidgetResponse" } },
          "owner_id": { "type": "string", "nullable": true },
          "is_public": { "type": "boolean" },
          "created_at": { "type": "string", "format": "date-time" },
          "updated_at": { "type": "string", "format": "date-time" }
        }
      },
      "DashboardListResponse": {
        "allOf": [
          { "$ref": "#/components/schemas/PaginatedResponse" },
          {
            "type": "object",
            "properties": {
              "items": { "type": "array", "items": { "$ref": "#/components/schemas/DashboardResponse" } }
            }
          }
        ]
      },
...
    }
  },
  "paths": {
...
    "/api/v1/dashboards": {
      "post": {
        "operationId": "create_dashboard",
        "summary": "Create dashboard",
        "description": "Creates a new dashboard with optional widgets.",
        "tags": ["Dashboards"],
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DashboardCreateRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Dashboard created",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "422": {
            "description": "Invalid dashboard data",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "list_dashboards",
        "summary": "List dashboards",
        "description": "Returns a paginated list of dashboards owned by the current user.",
        "tags": ["Dashboards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "page", "in": "query", "schema": { "type": "integer", "minimum": 1, "default": 1 } },
          { "name": "page_size", "in": "query", "schema": { "type": "integer", "minimum": 1, "maximum": 100, "default": 20 } }
        ],
        "responses": {
          "200": {
            "description": "List of dashboards",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardListResponse" }
              }
            }
          }
        }
      }
    },
    "/api/v1/dashboards/{dashboard_id}": {
      "get": {
        "operationId": "get_dashboard",
        "summary": "Get dashboard by ID",
        "description": "Returns a single dashboard with all its widgets.",
        "tags": ["Dashboards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "dashboard_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "Dashboard found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "404": {
            "description": "Dashboard not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "put": {
        "operationId": "update_dashboard",
        "summary": "Update dashboard",
        "description": "Updates an existing dashboard's properties and widgets.",
        "tags": ["Dashboards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "dashboard_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/DashboardUpdateRequest" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Dashboard updated",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DashboardResponse" }
              }
            }
          },
          "404": {
            "description": "Dashboard not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "delete_dashboard",
        "summary": "Delete dashboard",
        "description": "Soft-deletes a dashboard and its widgets.",
        "tags": ["Dashboards"],
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          { "name": "dashboard_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "204": {
            "description": "Dashboard deleted"
          },
          "404": {
            "description": "Dashboard not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ErrorResponse" }
              }
            }
          }
        }
      }
    },
...
    "/ws/dashboard/{dashboard_id}": {
      "parameters": [
        { "name": "dashboard_id", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
      ],
      "get": {
        "operationId": "dashboard_websocket",
        "summary": "Dashboard real-time WebSocket",
        "description": "Establishes a WebSocket connection for real-time dashboard metric updates. Supports subscribe, refresh, and ping messages.",
        "tags": ["WebSocket"],
        "parameters": [
          { "name": "token", "in": "query", "schema": { "type": "string" }, "description": "Optional JWT token for authentication" }
        ],
        "responses": {
          "101": {
            "description": "WebSocket connection established"
          },
          "426": {
            "description": "Upgrade required"
          }
        }
      }
    }
  }
}